Comments Added to Code

+++ core-csharp-practice/gcr-codebase/programming-elements/CelsiusToFah.cs
@@ -2,8 +2,13 @@ using System;
 
 class CeltoFah{
 		static void Main(){
-		float c=37.5;
-		float f=(c*(9/5)) + 32;
+			
+		// Temperature in Celsius
+		double c=37.5;
+		
+		// Convert Celsius to Fahrenheit using formula:
+        // Fahrenheit = (Celsius × 9 / 5) + 32
+		double f=(c*(9/5)) + 32;
 		Console.WriteLine(f);
 		}
 	}
